Split up name.mjs
