+++ src/fn/index.mjs
@@ -1,5 +1,15 @@
+export * from './get-encoding.mjs';
 export * from './get-font-file-data.mjs';
+export * from './get-name-by-id.mjs';
 export * from './get-unicode-range.mjs';
+export * from './mac-languages.mjs';
+export * from './make-name-table.mjs';
 export * from './make-os2-table.mjs';
+export * from './name-table-names.mjs';
+export * from './parse-name-table.mjs';
+export * from './parse-opentype-table-entries.mjs';
 export * from './parse-os2-table.mjs';
+export * from './parse-woff-table-entries.mjs';
+export * from './platforms.mjs';
 export * from './uncompress-table.mjs';
+export * from './windows-languages.mjs';
