+++ src/fn/platforms.mjs
@@ -1,0 +1,6 @@
+export const platforms = {
+    0: 'unicode',
+    1: 'macintosh',
+    2: 'reserved',
+    3: 'windows'
+};
