+++ src/tables/sfnt.mjs
@@ -14,7 +14,6 @@ import hhea from './hhea.mjs';
 import hmtx from './hmtx.mjs';
 import ltag from './ltag.mjs';
 import maxp from './maxp.mjs';
-import _name from './name.mjs';
 import post from './post.mjs';
 import gsub from './gsub.mjs';
 import meta from './meta.mjs';
@@ -27,7 +26,7 @@ import cvar from './cvar.mjs';
 import gvar from './gvar.mjs';
 import gasp from './gasp.mjs';
 import svg from './svg.mjs';
-import { getUnicodeRange, makeOS2Table } from '../fn/index.mjs';
+import { getUnicodeRange, makeOS2Table, makeNameTable } from '../fn/index.mjs';
 
 function log2(v) {
     return Math.log(v) / Math.log(2) | 0;
@@ -52,20 +51,20 @@ function computeCheckSum(bytes) {
 
 function makeTableRecord(tag, checkSum, offset, length) {
     return new table.Record('Table Record', [
-        {name: 'tag', type: 'TAG', value: tag !== undefined ? tag : ''},
-        {name: 'checkSum', type: 'ULONG', value: checkSum !== undefined ? checkSum : 0},
-        {name: 'offset', type: 'ULONG', value: offset !== undefined ? offset : 0},
-        {name: 'length', type: 'ULONG', value: length !== undefined ? length : 0}
+        { name: 'tag', type: 'TAG', value: tag !== undefined ? tag : '' },
+        { name: 'checkSum', type: 'ULONG', value: checkSum !== undefined ? checkSum : 0 },
+        { name: 'offset', type: 'ULONG', value: offset !== undefined ? offset : 0 },
+        { name: 'length', type: 'ULONG', value: length !== undefined ? length : 0 }
     ]);
 }
 
 function makeSfntTable(tables) {
     const sfnt = new table.Table('sfnt', [
-        {name: 'version', type: 'TAG', value: 'OTTO'},
-        {name: 'numTables', type: 'USHORT', value: 0},
-        {name: 'searchRange', type: 'USHORT', value: 0},
-        {name: 'entrySelector', type: 'USHORT', value: 0},
-        {name: 'rangeShift', type: 'USHORT', value: 0}
+        { name: 'version', type: 'TAG', value: 'OTTO' },
+        { name: 'numTables', type: 'USHORT', value: 0 },
+        { name: 'searchRange', type: 'USHORT', value: 0 },
+        { name: 'entrySelector', type: 'USHORT', value: 0 },
+        { name: 'rangeShift', type: 'USHORT', value: 0 }
     ]);
     sfnt.tables = tables;
     sfnt.numTables = tables.length;
@@ -80,7 +79,7 @@ function makeSfntTable(tables) {
     let offset = sfnt.sizeOf() + (makeTableRecord().sizeOf() * sfnt.numTables);
     while (offset % 4 !== 0) {
         offset += 1;
-        tableFields.push({name: 'padding', type: 'BYTE', value: 0});
+        tableFields.push({ name: 'padding', type: 'BYTE', value: 0 });
     }
 
     for (let i = 0; i < tables.length; i += 1) {
@@ -88,18 +87,18 @@ function makeSfntTable(tables) {
         check.argument(t.tableName.length === 4, 'Table name' + t.tableName + ' is invalid.');
         const tableLength = t.sizeOf();
         const tableRecord = makeTableRecord(t.tableName, computeCheckSum(t.encode()), offset, tableLength);
-        recordFields.push({name: tableRecord.tag + ' Table Record', type: 'RECORD', value: tableRecord});
-        tableFields.push({name: t.tableName + ' table', type: 'RECORD', value: t});
+        recordFields.push({ name: tableRecord.tag + ' Table Record', type: 'RECORD', value: tableRecord });
+        tableFields.push({ name: t.tableName + ' table', type: 'RECORD', value: t });
         offset += tableLength;
         check.argument(!isNaN(offset), 'Something went wrong calculating the offset.');
         while (offset % 4 !== 0) {
             offset += 1;
-            tableFields.push({name: 'padding', type: 'BYTE', value: 0});
+            tableFields.push({ name: 'padding', type: 'BYTE', value: 0 });
         }
     }
 
     // Table records need to be sorted alphabetically.
-    recordFields.sort(function(r1, r2) {
+    recordFields.sort(function (r1, r2) {
         if (r1.value.tag > r2.value.tag) {
             return 1;
         } else {
@@ -217,7 +216,7 @@ function fontToSfntTable(font) {
     }
     if (font.italicAngle < 0) {
         macStyle |= font.macStyleValues.ITALIC;
-    } 
+    }
 
     const headTable = head.make({
         flags: 3, // 00000011 (baseline for font at y=0; left sidebearing point at x=0)
@@ -261,7 +260,7 @@ function fontToSfntTable(font) {
         usWinAscent: globals.yMax,
         usWinDescent: Math.abs(globals.yMin),
         ulCodePageRange1: 1, // FIXME: hard-code Latin 1 support for now
-        sxHeight: metricsForChar(font, 'xyvw', {yMax: Math.round(globals.ascender / 2)}).yMax,
+        sxHeight: metricsForChar(font, 'xyvw', { yMax: Math.round(globals.ascender / 2) }).yMax,
         sCapHeight: metricsForChar(font, 'HIKLEFJMNTZBDPRAGOQSUVWXY', globals).yMax,
         usDefaultChar: font.hasChar(' ') ? 32 : 0, // Use space as the default character, if available.
         usBreakChar: font.hasChar(' ') ? 32 : 0, // Use space as the break character, if available.
@@ -302,7 +301,7 @@ function fontToSfntTable(font) {
         }
 
         if (!names[platform].postScriptName) {
-            names[platform].postScriptName = {en: postScriptName};
+            names[platform].postScriptName = { en: postScriptName };
         }
     }
 
@@ -332,7 +331,7 @@ function fontToSfntTable(font) {
     }
 
     const languageTags = [];
-    const nameTable = _name.make(names, languageTags);
+    const nameTable = makeNameTable(names, languageTags);
     const ltagTable = (languageTags.length > 0 ? ltag.make(languageTags) : undefined);
 
     const postTable = post.make(font);
@@ -363,7 +362,7 @@ function fontToSfntTable(font) {
         stat,
         avar,
         cvar,
-        fvar, 
+        fvar,
         gvar,
         gasp,
         svg,
